# use IndexList to improve performance of dot op