<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] --> Performance optimization 

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] --> OPs

### Describe
<!-- Describe what this PR does --> 
 Using IndexList instead of arrays of indices can speed up CPU and GPU performance. This PR use it to improve performance of dot op. For more information, please refer to #25132

### performace
   **GPU**: v100, cuda10
  |op|input shape|before|after| speed up |
  |---|---|---|---|---|
  |dot|[1000, 1000]|0.210704 ms|0.125058 ms|1.7x| 

**由于修改错误，以上数据存在问题，实际没有提升，此PR暂时关闭**